Drop support for --all in favor of explicit architecture arguments (#265)
With #263 removing support for 3 architectures, --all now builds the
images only for two remaining supported architectures and may silently
fail if add-on developers rely on it for building images for the
deprecated ones. To make the CI failing in such case, deprecate this
option and suggest to use explicit arguments for the built architectures
(can be combined).